splits qpid into header and cpp
